\name{btc}
\alias{btc}
\docType{data}
\title{Bitcoin and US stock indices data}
\description{Daily close exchange rate of Bitcoin and of three composite indices of the US stock market from 17 September 2014 to 30 September 2020.}
\usage{data(btc)}
\format{
  A data.frame with a total of 1521 observations on the following 5 variables:
  \describe{
    \item{\code{Date}}{Day of observation.}
    \item{\code{DJA}}{Exchange rate of Dow Jones Average index.}
    \item{\code{IXIC}}{Exchange rate of Nasdaq Composite index.}
    \item{\code{GSPC}}{Exchange rate of Standard\&Poor 500 index.}
    \item{\code{BTC}}{Exchange rate of Bitcoin.}
  }
}
%\references{}
